refactor(api): move controller
to identity-access-management bounded context
